+++ CHANGES.rst
@@ -24,6 +24,11 @@ New features:
 
 - Add fields in field editing UI to the current selected fieldset.
   [thet]
+- Allow specifying a vocabulary in the form of *key|label* in (multi)choice fields
+  [tomgross]
+
+- Removed very old import conditions
+  [tomgross]
 
 Bug fixes: